backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
yandex-music==2.2.0
requests==2.31.0
urllib3>=1.26.0
supabase==2.0.3
python-dotenv==1.0.0
httpx==0.24.1
